Mostly misspelling fixes
No changes affect code as the only thing that wasn't in comments was changing the local variables of "producableFrac" to "producibleFrac" (and I honestly can't blame anyone for not knowing that irregular spelling). All this does is fixes spelling in several cases (I'm 99.99% sure that initiatizes is supposed to be initializes not initiates), adds apostrophes in a few contractions, and changes a mistyped comment of "'sCost'roduct.pCost = sCost" to "'sCost' = product.pCost".